refactor: fixed type in ui-entities package
