+++ src/renderer/entities/governance/ui/VoteCharts/ReferendumVoteChart/ReferendumVoteChart.tsx
@@ -1,7 +1,7 @@
 import { useI18n } from '@/app/providers';
 import { Tooltip } from '@/shared/ui/Popovers';
 import { FootnoteText, HelpText } from '@/shared/ui/Typography';
-import { VoteChart } from '@shared/ui-entries';
+import { VoteChart } from '@shared/ui-entities';
 
 type Props = {
   aye: number;
