+++ src/renderer/shared/ui-entities/VoteChart/VoteChart.stories.tsx
