+++ src/renderer/shared/ui-entities/VoteChart/VoteChart.tsx
