+++ src/renderer/shared/ui-entities/index.ts
