+++ src/renderer/shared/ui-entities/tsconfig.json
